+++ app/root.tsx
@@ -159,7 +159,8 @@ export default function App() {
           src="https://analytics.sthom.kiwi/umami.js"
         ></script>
       </head>
-      <body className="h-full bg-bg">
+      <body className="h-full bg-bg text-text">
+        <script src="/themes.js"></script>
         <InstallContextProvider>
           <Outlet />
         </InstallContextProvider>